# Majority vote improvement